PYCBC-458: Fix Clang and Python warnings during installation
Motivation
----------
There are various Clang compiler warnings that come up during
installation, which should be addresed. Additionally, the
acouchbase iterator module is included in Python<3.4 packages,
but this module relies on the 'yield from' syntax which is only
available on Python 3.3 and above, and asyncio which is only
available in Python 3.4 and above. We may backport this to
Python2.x using Trollius and Trollius.From in future, but
for now this module should not be included in Python<3.4
packages.

Modifications
-------------
- Fixed various Clang warnings.
- Change Ping code to only fill the 'details' field when an
  error has occurred, matching the behaviour of the JSON
  code from LCB. Also use the svc->rc field rather than
  svc->status, as for LCB. Adjust JSON schema in
  tests to allow for optional 'details' field.
- Selective inclusion of acouchbase.iterator module for
  Python>=3.4
Results
-------
All known warnings/errors eradicated. acouchbase package still
present in Python>=3.4, with all tests passing.

Reviewed-on: http://review.couchbase.org/89347
Tested-by: Build Bot <build@couchbase.com>

+++ src/opresult.c
@@ -303,7 +303,7 @@ pycbc_sdresult_addresult(pycbc__SDResult *obj, size_t ii, PyObject *item)
     if (obj->results == NULL) {
         obj->results = PyList_New(PyTuple_GET_SIZE(obj->specs));
     }
-    pycbc_assert(ii < PyTuple_GET_SIZE(obj->specs));
+    pycbc_assert(ii < (size_t)PyTuple_GET_SIZE(obj->specs));
     PyList_SetItem(obj->results, ii, item);
     Py_INCREF(item); /* To normalize refcount semantics */
 }
